democratos/requirements.txt
Django==1.11.7
django-ckeditor==5.2.2
django-comments-xtd==2.0.2
django-contrib-comments==1.8.0
django-registration-redux==1.6
django-render-block==0.5
django-simple-captcha==0.5.5
djangoajax==2.3.7
djangorestframework==3.6.4
docutils==0.14
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
